react-native example: rename app name to reactNative
