+++ examples/sdk/reactNative/android/app/src/main/java/com/reactnative/MainApplication.kt
@@ -1,4 +1,4 @@
-package com.rnbtexample
+package com.reactnative
 
 import android.app.Application
 import com.facebook.react.PackageList
