+++ examples/sdk/reactNative/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">rnbtexample</string>
+    <string name="app_name">reactNative</string>
 </resources>
