+++ examples/sdk/reactNative/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "rnbtexample",
-  "displayName": "rnbtexample"
+  "name": "reactNative",
+  "displayName": "reactNative"
 }
