+++ examples/sdk/reactNative/ios/Podfile
@@ -14,7 +14,7 @@ if linkage != nil
   use_frameworks! :linkage => linkage.to_sym
 end
 
-target 'rnbtexample' do
+target 'reactNative' do
   config = use_native_modules!
 
   use_react_native!(
@@ -23,7 +23,7 @@ target 'rnbtexample' do
     :app_path => "#{Pod::Config.instance.installation_root}/.."
   )
 
-  target 'rnbtexampleTests' do
+  target 'reactNativeTests' do
     inherit! :complete
     # Pods for testing
   end
