+++ examples/sdk/reactNative/ios/reactNative.xcodeproj/xcshareddata/xcschemes/reactNative.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "rnbtexample.app"
-               BlueprintName = "rnbtexample"
-               ReferencedContainer = "container:rnbtexample.xcodeproj">
+               BuildableName = "reactNative.app"
+               BlueprintName = "reactNative"
+               ReferencedContainer = "container:reactNative.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "rnbtexampleTests.xctest"
-               BlueprintName = "rnbtexampleTests"
-               ReferencedContainer = "container:rnbtexample.xcodeproj">
+               BuildableName = "reactNativeTests.xctest"
+               BlueprintName = "reactNativeTests"
+               ReferencedContainer = "container:reactNative.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "rnbtexample.app"
-            BlueprintName = "rnbtexample"
-            ReferencedContainer = "container:rnbtexample.xcodeproj">
+            BuildableName = "reactNative.app"
+            BlueprintName = "reactNative"
+            ReferencedContainer = "container:reactNative.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "rnbtexample.app"
-            BlueprintName = "rnbtexample"
-            ReferencedContainer = "container:rnbtexample.xcodeproj">
+            BuildableName = "reactNative.app"
+            BlueprintName = "reactNative"
+            ReferencedContainer = "container:reactNative.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
