+++ examples/sdk/reactNative/ios/reactNative/AppDelegate.h
