+++ examples/sdk/reactNative/ios/reactNative/Images.xcassets/AppIcon.appiconset/Contents.json
