+++ examples/sdk/reactNative/ios/reactNative/Images.xcassets/Contents.json
