+++ examples/sdk/reactNative/ios/reactNative/PrivacyInfo.xcprivacy
