+++ examples/sdk/reactNative/ios/reactNativeTests/Info.plist
