+++ examples/sdk/reactNative/ios/reactNativeTests/reactNativeTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface rnbtexampleTests : XCTestCase
+@interface reactNativeTests : XCTestCase
 
 @end
 
-@implementation rnbtexampleTests
+@implementation reactNativeTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL (^)(UIView *view))test
 {
