+++ examples/sdk/reactNative/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "rnbtexample",
+    "name": "reactNative",
     "version": "0.0.1",
     "private": true,
     "scripts": {
